Thêm enum material để khi tạo form thêm điểm cho người dùng nó dễ hơn Tạo luôn quy định luôn cho dễ

+++ src/main/java/com/example/EcoTS/Models/Points.java
@@ -44,8 +44,8 @@ public class Points {
     @OneToOne
     @JoinColumn(name = "user_id", referencedColumnName = "id")
     private Users user;
-    private Long saveCo2;
-    private Long totalTrashCollect;
+    private double saveCo2;
+    private double totalTrashCollect;
     @CreationTimestamp
     @Column(name = "created_at", updatable = false)
     private Timestamp createdAt;
